requirements.txt
annotated-types==0.7.0
anyio==4.4.0
arq==0.26.0
async-timeout==4.0.3
boto3==1.34.153
botocore==1.34.153
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.112.0
h11==0.14.0
hiredis==3.0.0
idna==3.7
jmespath==1.0.1
scikit-image==0.24.0
numpy==2.1.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
redis==4.6.0
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
